ArticleService modified

+++ src/main/java/ml/sansejin/sancolor/entity/Article.java
@@ -29,15 +29,15 @@ public class Article {
         this.user_name = user_name;
     }
 
-    public Integer getUser_id() {
+    public Long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(Integer user_id) {
+    public void setUser_id(Long user_id) {
         this.user_id = user_id;
     }
 
-    private Integer user_id;
+    private Long user_id;
 
     public Long getId() {
         return id;
